// projects/frontdesk/src/app/vvtk-core/interfaces/custom-route.ts
export interface CustomRoute {
  path: string;
  referPath: string;
  customPath: string;
  imgPcSrc?: string;
  imgTabletSrc?: string;
  imgMobileSrc?: string;
  title?: string;
  meta?: string;
  jsonLd?: string;
  imgPcAlt?: string;
  imgPcTitle?: string;
  imgTabletAlt?: string;
  imgTabletTitle?: string;
  imgMobileAlt?: string;
  imgMobileTitle?: string;
}